Restructure code and modules (#25)
* Rename module containing api client

* Move coordinator to separate module

* Improve coordinator type hinting

* Compact data dictionary generation

* Remove unused function
