+++ custom_components/tdarr/sensor.py
@@ -14,6 +14,7 @@ from . import (
     TdarrLibraryEntity,
     TdarrNodeEntity,
 )
+from .coordinator import TdarrDataUpdateCoordinator
 from .const import (
     DOMAIN,
     COORDINATOR,
@@ -227,7 +228,7 @@ async def async_setup_entry(hass, config_entry, async_add_entities):
 
 class TdarrServerSensor(TdarrServerEntity, SensorEntity):
 
-    def __init__(self, coordinator, options, entity_description: TdarrSensorEntityDescription):
+    def __init__(self, coordinator: TdarrDataUpdateCoordinator, options, entity_description: TdarrSensorEntityDescription):
         _LOGGER.info("Creating server level sensor %s", entity_description.key)
         super().__init__(coordinator, entity_description)
 
@@ -247,7 +248,7 @@ class TdarrServerSensor(TdarrServerEntity, SensorEntity):
 
 class TdarrLibrarySensor(TdarrLibraryEntity, SensorEntity):
 
-    def __init__(self, coordinator, library_id, options, entity_description: TdarrSensorEntityDescription):
+    def __init__(self, coordinator: TdarrDataUpdateCoordinator, library_id: str, options, entity_description: TdarrSensorEntityDescription):
         _LOGGER.info("Creating library %s level sensor %s", library_id, entity_description.key)
         super().__init__(coordinator, library_id, entity_description)
 
@@ -278,7 +279,7 @@ class TdarrLibrarySensor(TdarrLibraryEntity, SensorEntity):
 
 class TdarrNodeSensor(TdarrNodeEntity, SensorEntity):
 
-    def __init__(self, coordinator, node_id, options, entity_description: TdarrSensorEntityDescription):
+    def __init__(self, coordinator: TdarrDataUpdateCoordinator, node_id: str, options, entity_description: TdarrSensorEntityDescription):
         _LOGGER.info("Creating node %s level sensor %s", node_id, entity_description.key)
         super().__init__(coordinator, node_id, entity_description)
